Add asset-root option to support serving out of subdir...
Fixes #4.

The path of the source file was being concatenated onto the asset path,
which is incorrect when the root dir is common. For example, `full-asset-path`
returned the path `public/css/img/background.png`, later failing when the
references were replaced.

The change I've introduced is breaking, since now all templated references are
relative to the asset root. In the example directory, the reference
`${nested.js}` inside `nested/index.html` is now `${nested/nested.js}`. I think
this makes more sense because assets are typically served from a single
directory; it's far easier to remember the asset root than the path of the asset
relative to the source file. There may, however, be something I've not
considered - happy to make changes if necessary.